+++ ChangeLog.md
@@ -22,6 +22,7 @@
 * Support d'Invocation d'un démon (nécromancien)
 
 ### Autres améliorations
+* Prise en compte de l'option --secret pour !cof-bonus-couvert
 * Armes avec bonus de DEF
 * Possibilité d'afficher un message en cas de coup critique reçu.
 * Ajout d'un effet temporaire effetRetarde